Sneak in a temporary repo for Forge-based Gridware installations.
This is horribly inefficient - for every package to be installed thus,
we clone the relevant upstream Gridware repo (`main` or `volatile`) and
inject it as an additional available repo to Gridware during the
installation.

We could probably reduce this to only when the repository is otherwise
unavailable - but that's for the `install.sh` scripts, not here.