Always render depends files on import.
As the repository that the package was exported from may not match the
repository which the package was imported from the current hard-coded nature of
the repository in the depends files can render packages broken when they're
imported.

This change modifies the importer such that the depends files are always
rendered on import.